+++ src/main/java/com/foxminded/ums/controllers/TeachersRestController.java
@@ -1,8 +1,8 @@
 package com.foxminded.ums.controllers;
 
-import com.foxminded.ums.dto.StudentDto;
 import com.foxminded.ums.dto.TeacherDto;
 import com.foxminded.ums.exeptions.ErrorResponce;
+import com.foxminded.ums.security.SecurityHelper;
 import com.foxminded.ums.service.TeacherService;
 import com.foxminded.ums.validation.UUID;
 import io.swagger.v3.oas.annotations.Operation;
@@ -18,6 +18,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -40,6 +42,9 @@ public class TeachersRestController {
     @Autowired
     private TeacherService teacherService;
 
+    @Autowired
+    private SecurityHelper securityHelper;
+
     @Operation(summary = "Show List of Teachers page by page",
             description = "Show One Page of List of Teachers",
             tags = {"teachers"})
@@ -55,10 +60,15 @@ public class TeachersRestController {
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<List<TeacherDto>> findTeachers(
             @Parameter(description = "page > 0 (default = 0), size > 1 (default = 5); masked by default values")
-            @PageableDefault(page = 0, size = 5) Pageable pageable) {
-        List<TeacherDto> teacherDtos = teacherService.findTeachersPageable(pageable);
+            @PageableDefault(page = 0, size = 5) Pageable pageable,
+            Authentication authentication) {
+        if (securityHelper.isAdmin(authentication)) {
+            List<TeacherDto> teacherDtos = teacherService.findTeachersPageable(pageable);
 
-        return ResponseEntity.ok().body(teacherDtos);
+            return ResponseEntity.ok().body(teacherDtos);
+        }
+
+        throw new AccessDeniedException("Endpoint allowed only for Admins");
     }
 
     @Operation(summary = "Add new Teacher",
@@ -77,10 +87,15 @@ public class TeachersRestController {
     public ResponseEntity<TeacherDto> addTeacher(
             @Parameter(description = "Student to add. Cannot null or empty",
                     required = true, schema = @Schema(implementation = TeacherDto.class))
-            @Valid @RequestBody TeacherDto teacherDto) {
-        TeacherDto addedTeacher = teacherService.addTeacher(teacherDto);
+            @Valid @RequestBody TeacherDto teacherDto,
+            Authentication authentication) {
+        if (securityHelper.isAdmin(authentication)) {
+            TeacherDto addedTeacher = teacherService.addTeacher(teacherDto);
 
-        return ResponseEntity.status(HttpStatus.CREATED).body(addedTeacher);
+            return ResponseEntity.status(HttpStatus.CREATED).body(addedTeacher);
+        }
+
+        throw new AccessDeniedException("Endpoint allowed only for Admins");
     }
 
     @Operation(summary = "Find Teacher by ID",
@@ -99,12 +114,17 @@ public class TeachersRestController {
     @ResponseStatus(HttpStatus.OK)
     public ResponseEntity<TeacherDto> findTeacher(
             @Parameter(description = "Teacher UUID", required = true)
-            @Valid @PathVariable("id") @UUID String id) {
-        java.util.UUID teacherId = java.util.UUID.fromString(id);
+            @Valid @PathVariable("id") @UUID String id,
+            Authentication authentication) {
+        if (securityHelper.isAdmin(authentication)) {
+            java.util.UUID teacherId = java.util.UUID.fromString(id);
 
-        TeacherDto teacherDto = teacherService.findTeacher(teacherId);
+            TeacherDto teacherDto = teacherService.findTeacher(teacherId);
 
-        return ResponseEntity.ok().body(teacherDto);
+            return ResponseEntity.ok().body(teacherDto);
+        }
+
+        throw new AccessDeniedException("Endpoint allowed only for Admins");
     }
 
     @Operation(summary = "Update existed Teacher",
@@ -129,13 +149,18 @@ public class TeachersRestController {
             @Parameter(description = "Teacher to update. Cannot null or empty", required = true)
             @Valid @RequestBody TeacherDto teacherDto,
             @Parameter(description = "Teacher UUID", required = true)
-            @Valid @PathVariable("id") @UUID String id) {
-        java.util.UUID teacherId = java.util.UUID.fromString(id);
-        teacherDto.setId(teacherId);
+            @Valid @PathVariable("id") @UUID String id,
+            Authentication authentication) {
+        if (securityHelper.isAdmin(authentication)) {
+            java.util.UUID teacherId = java.util.UUID.fromString(id);
+            teacherDto.setId(teacherId);
 
-        TeacherDto updatedTeacher = teacherService.updateTeacher(teacherDto);
+            TeacherDto updatedTeacher = teacherService.updateTeacher(teacherDto);
 
-        return ResponseEntity.ok().body(updatedTeacher);
+            return ResponseEntity.ok().body(updatedTeacher);
+        }
+
+        throw new AccessDeniedException("Endpoint allowed only for Admins");
     }
 
     @Operation(summary = "Delete existed Student",
@@ -155,11 +180,16 @@ public class TeachersRestController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public ResponseEntity<TeacherDto> deleteTeacher(
             @Parameter(description = "Teacher to delete. Cannot null or empty", required = true)
-            @Valid @PathVariable("id") @UUID String id) {
-        java.util.UUID teacherId = java.util.UUID.fromString(id);
-        teacherService.deleteTeacher(teacherId);
+            @Valid @PathVariable("id") @UUID String id,
+            Authentication authentication) {
+        if (securityHelper.isAdmin(authentication)) {
+            java.util.UUID teacherId = java.util.UUID.fromString(id);
+            teacherService.deleteTeacher(teacherId);
 
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
+        }
+
+        throw new AccessDeniedException("Endpoint allowed only for Admins");
     }
 
 }
